requirements.txt
Django==4.1.4
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
django-safedelete==1.3.1